% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogate.R
\name{autogate}
\alias{autogate}
\title{autogate}
\usage{
autogate(
  param.dir,
  data.dir,
  transform.filename,
  fcs.experiment,
  fcs.panel,
  agp,
  fcs.gates.all,
  compensate = FALSE,
  compensation.filename,
  setup = FALSE
)
}
\description{
autogate
}
