% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_directories.R
\name{create.directories}
\alias{create.directories}
\title{create.directories}
\usage{
create.directories(agp, fcs.experiment)
}
\arguments{
\item{agp}{The AutoGating parameter list.}
}
\value{
List of the output figure directories
}
\description{
Creates the output folders for the pipeline.
}
