% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_activation_gates_by_thresholds.R
\name{do.activation.gates.by.thresholds}
\alias{do.activation.gates.by.thresholds}
\title{do.activation.gates.by.thresholds}
\usage{
do.activation.gates.by.thresholds(
  fcs.expr.data,
  fcs.population.gates,
  fcs.figure.dir,
  fcs.activation.thresholds,
  fcs.activation.gates,
  fcs.activation.gates.calculate,
  fcs.activation.gates.param
)
}
\arguments{
\item{fcs.expr.data}{The marker expression data from the fcs files.}

\item{fcs.population.gates}{List of population-defining (cell type) gates.}

\item{fcs.figure.dir}{Directory where the figures will be generated.}

\item{fcs.activation.thresholds}{The thresholds used to define the positive/negative
cutoffs for the activation marker gates.}

\item{fcs.activation.gates.calculate}{Logical?}

\item{fcs.activation.gates.param}{Parameters for calculating the activation gates.}
}
\value{
description
}
\description{
A function to calculate and execute gating for activation markers in the panel.
}
