% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_activation_gates_from_definition.R
\name{do.activation.gates.def}
\alias{do.activation.gates.def}
\title{do.activation.gates.def}
\usage{
do.activation.gates.def(
  fcs.gates,
  fcs.gates.data,
  fcs.activ.gates.def,
  calculate.gates,
  figure.actgate.dir
)
}
\arguments{
\item{fcs.gates}{The gates for flow cytometry data.}

\item{fcs.gates.data}{Data used for generating? gates.}

\item{fcs.activ.gates.def}{Definitions of the activation marker gates.}

\item{calculate.gates}{Logical}

\item{figure.actgate.dir}{Directory for output?}
}
\value{
fcs.gates
}
\description{
A function to calculate and execute gating for activation markers in the panel.
}
