% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_activation_gates.R
\name{do.activation.gates}
\alias{do.activation.gates}
\title{do.activation.gates}
\usage{
do.activation.gates(
  fcs.gates,
  fcs.gates.data,
  fcs.activ.gates.definition,
  calculate.gates,
  figure.actgate.dir
)
}
\arguments{
\item{fcs.expr.data}{The marker expression data from the fcs files.}

\item{fcs.population.gates}{List of population-defining (cell type) gates.}

\item{fcs.figure.dir}{Directory where the figures will be generated.}

\item{fcs.activation.thresholds}{The thresholds used to define the positive/negative
cutoffs for the activation marker gates.}

\item{fcs.activation.gates.calculate}{Logical?}

\item{fcs.activation.gates.param}{Parameters for calculating the activation gates.}
}
\value{
description
}
\description{
A function to calculate and execute gating for activation markers in the panel.
}
