% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_gate_generic.R
\name{do.gate.generic}
\alias{do.gate.generic}
\title{do.gate.generic}
\usage{
do.gate.generic(
  fcs.gates,
  fcs.gate.data,
  fcs.gate.def,
  calculate.gate,
  figure.dir
)
}
\arguments{
\item{fcs.gates}{The gates for flow cytometry data.}

\item{fcs.gate.def}{Definitions of the gates.}

\item{figure.dir}{Directory for output?}

\item{fcs.gates.data}{Data used for generating? gates.}

\item{calculate.gates}{Logical?}
}
\value{
fcs.gates
}
\description{
The generic gating function. Will generate flow cytometry gates based on the
gate structure and data fed in. Outputs plots and returns gates.
}
