% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_population_gates.R
\name{do.population.gates}
\alias{do.population.gates}
\title{do.population.gates}
\usage{
do.population.gates(
  fcs.gates,
  fcs.gates.data,
  fcs.popul.gates.definition,
  calculate.gates,
  figure.popgate.dir
)
}
\arguments{
\item{fcs.gates}{The gates to be generated.}

\item{fcs.gates.data}{Data to be used to generate the gates.}

\item{fcs.popul.gates.definition}{Definitions of the gates, obtained from the
gate parameter table.}

\item{calculate.gates}{Logical?}

\item{figure.popgate.dir}{The directory where the figures will be generated.}
}
\value{
description
}
\description{
A function to calculate and execute gating for population (cell type) gates
in the panel.
}
