% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autogating_param_aurora.R
\name{get.autogating.param.aurora}
\alias{get.autogating.param.aurora}
\title{get.autogating.param.aurora}
\usage{
get.autogating.param.aurora(autogating.param)
}
\arguments{
\item{autogating.param}{The initial general parameter list provided by calling
get.autogating.param.minimal.}
}
\value{
autogating.param Returns updated autogating parameters
}
\description{
Get cytometer-specific parameters for Aurora files in order to perform autogating.
}
