% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autogating_param_id7000.R
\name{get.autogating.param.id7000}
\alias{get.autogating.param.id7000}
\title{get.autogating.param.id7000}
\usage{
get.autogating.param.id7000()
}
\arguments{
\item{autogating.param}{The initial general parameter list provided by calling
get.autogating.param.minimal.}
}
\value{
description
}
\description{
Get cytometer-specific parameters for ID7000 files in order to perform autogating.
}
