% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autogating_param_minimal.R
\name{get.autogating.param.minimal}
\alias{get.autogating.param.minimal}
\title{get.autogating.param.minimal}
\usage{
get.autogating.param.minimal()
}
\value{
The AutoGating parameter list.
}
\description{
Get a list of parameter essential for performing autogating.
To be called via get.autogating.param().
}
\examples{
agp <- get.autogating.param( cytometer = "Aurora" )

}
