% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autogating_param_symphony.R
\name{get.autogating.param.symphony}
\alias{get.autogating.param.symphony}
\title{get.autogating.param.symphony}
\usage{
get.autogating.param.symphony(autogating.param)
}
\arguments{
\item{autogating.param}{The initial general parameter list provided by calling
get.autogating.param.minimal.}
}
\value{
description
}
\description{
Get cytometer-specific parameters for FACSymphony files in order to perform
autogating.
}
