% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fcs_data.R
\name{read.fcs.data}
\alias{read.fcs.data}
\title{read.fcs.data}
\usage{
read.fcs.data(
  param.dir,
  data.dir,
  transform.filename,
  fcs.experiment,
  fcs.panel,
  agp,
  compensate = FALSE,
  compensation.filename,
  setup = TRUE
)
}
\arguments{
\item{param.dir}{File path to the transformation and compensation files.}

\item{data.dir}{File path to the fcs files.}

\item{transform.filename}{The csv file defining the parameters of the
biexponential transformation to be applied.}

\item{fcs.experiment}{Description of the cytometry experiment. Produced by
read.experiment().}

\item{fcs.panel}{Description of the cytometry panel. Produced by
read.panel.design().}

\item{agp}{AutoGating parameter list. Produced by get.autogating.param().}

\item{compensate}{Logical. If TRUE, compensation will be applied and a csv
file must be provided to compensation.filename.}

\item{compensation.filename}{A csv file providing the compensation matrix to
be applied. Required if compensate is set to TRUE.}

\item{setup}{Logical. If TRUE, uses only the files selected for defining
(calculating) the gates. If FALSE, processes all files.}
}
\value{
fcs.data The expression (time, scatter and marker) data from the fcs files.
}
\description{
Read in the fcs files for autogating.
}
