% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gate_definition.R
\name{read.gate.definition}
\alias{read.gate.definition}
\title{read.gate.definition}
\usage{
read.gate.definition(path, pop.gate.def.filename, act.gate.def.filename, agp)
}
\description{
read.gate.definition
}
