% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_panel_design.R
\name{read.panel.design}
\alias{read.panel.design}
\title{read.panel.design}
\usage{
read.panel.design(param.dir, fcs.panel.filename)
}
\arguments{
\item{param.dir}{File path to the fcs.panel.filename}

\item{fcs.panel.filename}{CSV file describing the panel structure.}
}
\value{
fcs.panel
}
\description{
Reads in the panel design csv file.
}
