% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_transformation.R
\name{read.transformation}
\alias{read.transformation}
\title{read.transformation}
\usage{
read.transformation(transformation.filepath, fcs.panel)
}
\arguments{
\item{transformation.filepath}{Path (including filename) to the location of
the file defining the transformation.}

\item{fcs.panel}{Description of the cytometry panel. Produced by read.panel.design().}
}
\value{
The transformation and inverse transformation lists for the channels.
}
\description{
Read in a csv file defining the biexponential transformation parameters.
}
