% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autogating_param_minimal.R
\name{get.autogating.param.minimal}
\alias{get.autogating.param.minimal}
\title{get.autogating.param.minimal}
\usage{
get.autogating.param.minimal()
}
\arguments{
\item{cytometer}{The cytometer used for acquiring the data. Supported options
are "Aurora" and "ID7000".}
}
\value{
description
}
\description{
Get a list of parameter essential for performing autogating.
}
